src/db/schemas/users.py
from pydantic import BaseModel


class UserSchema(BaseModel):
    user_id: int
    username: str
    password: str
    
    class Config:
        from_attributes = True
        

class UserCreateSchema(BaseModel):
    username: str
    password: str