pyproject.toml
[project]
name = "task-manager-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic (>=1.16.4,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[virtualenvs]
create = true
in-project = true